requirements.txt
pytesseract
pdf2image
Pillow
llama-cpp-python
tk